requirements.txt
requests>=2.28.0
psycopg2-binary>=2.9.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
torch>=2.0.0
pgvector>=0.2.0
flask
pyngrok
dropbox
dotenv